test(data-fabric): split entities integration tests into records/query/schema files
Divides the single 1886-line entities.integration.test.ts (~26 min, fully
serial) into three files so vitest runs them on parallel workers. Test bodies
are ported verbatim and both v0/v1 init modes are kept everywhere; measured
wall time drops to ~8.8 min.
