+++ tests/integration/shared/data-fabric/entities-records.integration.test.ts
@@ -9,13 +9,10 @@ import {
 import { registerResource } from '../../utils/cleanup';
 import { generateRandomString, generateRandomInt, generateRandomFloat, hasValidPagination } from '../../utils/helpers';
 import {
-  EntityAggregateFunction,
   EntityFieldDataType,
-  EntityHavingOperator,
   EntityRecord,
   FieldDisplayType,
   FieldMetaData,
-  JoinType,
   QueryFilterOperator,
   RawEntityGetResponse,
 } from '../../../../src/models/data-fabric/entities.types';
@@ -155,13 +152,12 @@ async function buildDummyRecord(entityMetadata: RawEntityGetResponse): Promise<R
 
 const modes: InitMode[] = ['v0', 'v1'];
 
-describe.each(modes)('Data Fabric Entities - Integration Tests [%s]', (mode) => {
+describe.each(modes)('Data Fabric Entities Records - Integration Tests [%s]', (mode) => {
   setupUnifiedTests(mode);
 
   let testEntityId: string | null = null;
   let entityMetadata: RawEntityGetResponse | null = null;
   const createdRecordIds: string[] = [];
-  const createdEntityIds: string[] = [];
 
   describe('getAll', () => {
     it('should retrieve all entities', async () => {
@@ -173,7 +169,10 @@ describe.each(modes)('Data Fabric Entities - Integration Tests [%s]', (mode) => 
       expect(Array.isArray(result)).toBe(true);
 
       if (result.length > 0) {
-        testEntityId = result[0].id;
+        // Prefer a stable entity for the fallback id — the entities-schema file
+        // runs on a parallel worker and can delete its transient sdk_* entities
+        // mid-run.
+        testEntityId = result.find((e) => !e.name.startsWith('sdk_'))?.id ?? result[0].id;
       }
     });
 
@@ -255,11 +254,17 @@ describe.each(modes)('Data Fabric Entities - Integration Tests [%s]', (mode) => 
       ]);
 
       expect(Array.isArray(allEntities)).toBe(true);
+
+      // The entities-schema file runs on a parallel worker and creates/deletes
+      // transient sdk_* entities; one deleted between the two list calls above
+      // would break the superset check, so compare stable entities only.
+      const stableTenantEntities = tenantEntities.filter((e) => !e.name.startsWith('sdk_'));
+
       // The combined list is a superset of the tenant-only list
-      expect(allEntities.length).toBeGreaterThanOrEqual(tenantEntities.length);
+      expect(allEntities.length).toBeGreaterThanOrEqual(stableTenantEntities.length);
 
       const allIds = new Set(allEntities.map((e) => e.id));
-      for (const tenantEntity of tenantEntities) {
+      for (const tenantEntity of stableTenantEntities) {
         expect(allIds.has(tenantEntity.id)).toBe(true);
       }
 
@@ -792,284 +797,6 @@ describe.each(modes)('Data Fabric Entities - Integration Tests [%s]', (mode) => 
     });
   });
 
-  describe('queryRecordsById', () => {
-    it('should query records with no filters', async () => {
-      const { entities } = getServices();
-      const config = getTestConfig();
-      const entityId = config.dataFabricTestEntityId || testEntityId;
-      if (!entityId) {
-        throw new Error('No entity ID available for testing');
-      }
-      const result = await entities.queryRecordsById(entityId);
-      expect(result).toBeDefined();
-      expect(result.items).toBeDefined();
-      expect(Array.isArray(result.items)).toBe(true);
-      expect(typeof result.totalCount).toBe('number');
-    });
-
-    it('should return paginated records when pageSize is provided', async () => {
-      const { entities } = getServices();
-      const config = getTestConfig();
-      const entityId = config.dataFabricTestEntityId || testEntityId;
-      if (!entityId) {
-        throw new Error('No entity ID available for testing');
-      }
-      const result = await entities.queryRecordsById(entityId, { pageSize: 2 });
-      expect(result).toBeDefined();
-      expect(Array.isArray(result.items)).toBe(true);
-      expect(result.items.length).toBeLessThanOrEqual(2);
-      expect(hasValidPagination(result)).toBe(true);
-    });
-
-    it('should return aggregate count when aggregates is provided without groupBy', async () => {
-      const { entities } = getServices();
-      const config = getTestConfig();
-      const entityId = config.dataFabricTestEntityId || testEntityId;
-      if (!entityId) {
-        throw new Error('No entity ID available for testing');
-      }
-      const result = await entities.queryRecordsById(entityId, {
-        aggregates: [
-          { function: EntityAggregateFunction.Count, field: 'Id', alias: 'total' },
-        ],
-      });
-      expect(result).toBeDefined();
-      expect(Array.isArray(result.items)).toBe(true);
-      expect(result.items.length).toBe(1);
-      const row = result.items[0] as Record<string, any>;
-      expect(row.total).toBeDefined();
-      expect(typeof row.total).toBe('number');
-      expect(row.total).toBeGreaterThanOrEqual(0);
-    });
-
-    // The tenant must have the `enable-having-on-query` feature flag; without it the
-    // server rejects havingFilter with a 400 naming the flag, which fails this test
-    // loudly rather than letting the coverage be silently absent.
-    it('should filter grouped results with havingFilter (HAVING)', async () => {
-      const { entities } = getServices();
-      const config = getTestConfig();
-      const entityId = config.dataFabricTestEntityId || testEntityId;
-      if (!entityId) {
-        throw new Error('No entity ID available for testing');
-      }
-      // Group by Id: always present on every entity, no fixture coupling — one
-      // group per record, each with cnt = 1, which keeps both assertions meaningful.
-      const base = {
-        selectedFields: ['Id'],
-        groupBy: ['Id'],
-        aggregates: [
-          { function: EntityAggregateFunction.Count, field: 'Id', alias: 'cnt' },
-        ],
-      };
-
-      // Every group has at least one record, so `cnt >= 1` must return every group.
-      const all = await entities.queryRecordsById(entityId, {
-        ...base,
-        havingFilter: {
-          aggregateFilters: [
-            { aggregateAlias: 'cnt', operator: EntityHavingOperator.GreaterThanOrEqual, value: '1' },
-          ],
-        },
-      });
-      expect(all.items.length).toBeGreaterThan(0);
-      all.items.forEach(item => {
-        expect((item as Record<string, unknown>).cnt).toBeGreaterThanOrEqual(1);
-      });
-
-      // An unsatisfiable threshold must return no groups. A backend that ignores
-      // havingFilter returns every group here — that is the failing signal.
-      const none = await entities.queryRecordsById(entityId, {
-        ...base,
-        havingFilter: {
-          aggregateFilters: [
-            { aggregateAlias: 'cnt', operator: EntityHavingOperator.GreaterThan, value: '1000000000' },
-          ],
-        },
-      });
-      expect(none.items).toHaveLength(0);
-    });
-
-    // Regression guard: DF reads `expansionLevel` only from the URL on POST record endpoints.
-    // If the SDK sends it in the body, DF silently ignores it and every level collapses to L0.
-    it('should expand reference fields at each expansionLevel (0-3)', async () => {
-      const { entities } = getServices();
-      const config = getTestConfig();
-      const entityId = config.dataFabricTestEntityId || testEntityId;
-      if (!entityId) {
-        throw new Error('No entity ID available for testing');
-      }
-
-      const levels = [0, 1, 2, 3] as const;
-      const responses = await Promise.all(
-        levels.map(level => entities.queryRecordsById(entityId, { expansionLevel: level, pageSize: 1 })),
-      );
-
-      responses.forEach((resp, i) => {
-        expect(resp, `expansionLevel=${levels[i]} returned no response`).toBeDefined();
-        expect(Array.isArray(resp.items), `expansionLevel=${levels[i]} items not an array`).toBe(true);
-      });
-
-      if (responses.some(r => r.items.length === 0)) {
-        throw new Error('Test entity has no records — expansionLevel diff cannot be verified. Insert at least one record into DATA_FABRIC_TEST_ENTITY_ID.');
-      }
-
-      const records = responses.map(r => r.items[0] as Record<string, any>);
-      const [l0Record, l1Record, l2Record, l3Record] = records;
-
-      // CreatedBy is a system reference field on every DF record.
-      // L0: raw GUID string. L1+: object envelope with Id.
-      expect(typeof l0Record.CreatedBy).toBe('string');
-
-      for (const [level, rec] of [[1, l1Record], [2, l2Record], [3, l3Record]] as const) {
-        expect(typeof rec.CreatedBy, `L${level} CreatedBy should be object`).toBe('object');
-        expect(rec.CreatedBy, `L${level} CreatedBy should not be null`).not.toBeNull();
-        expect(rec.CreatedBy, `L${level} CreatedBy should have Id`).toHaveProperty('Id');
-      }
-
-      // L2 inflates the L1 envelope into the referenced user's full record.
-      expect(Object.keys(l2Record.CreatedBy).length).toBeGreaterThan(
-        Object.keys(l1Record.CreatedBy).length,
-      );
-
-      // L3 must remain at least as expanded as L2 (deeper nesting is schema-dependent, but never shrinks).
-      expect(Object.keys(l3Record.CreatedBy).length).toBeGreaterThanOrEqual(
-        Object.keys(l2Record.CreatedBy).length,
-      );
-    });
-
-    // Multi-join. Requires the join fixture (a second, related entity, seeded
-    // with one record that matches a base record on the join key and one that
-    // matches nothing) provisioned in the test tenant and named via the
-    // DATA_FABRIC_TEST_JOIN_* env vars; throws when they are missing.
-    //
-    // These tests assert the join EFFECT, not just the response envelope: a
-    // backend that accepts the request but silently ignores the `joins` body
-    // key returns a perfectly normal envelope, so envelope-shape assertions
-    // alone cannot fail for the exact defect they exist to catch (this is how
-    // the original non-functional joins wiring shipped).
-    it('should return related-entity fields for a cross-entity LEFT join', async () => {
-      const { entities } = getServices();
-      const config = getTestConfig();
-      const entityId = config.dataFabricTestEntityId || testEntityId;
-      if (!entityId) {
-        throw new Error('No entity ID available for testing');
-      }
-      if (
-        !config.dataFabricTestJoinFieldName ||
-        !config.dataFabricTestJoinRelatedEntityName ||
-        !config.dataFabricTestJoinRelatedFieldName
-      ) {
-        throw new Error('DATA_FABRIC_TEST_JOIN_* env vars are required for the join test');
-      }
-      const relatedEntity = config.dataFabricTestJoinRelatedEntityName;
-
-      const result = await entities.queryRecordsById(entityId, {
-        // Join queries require a projection; select the two join keys — the
-        // base-entity key unqualified (non-ambiguous), the related key qualified.
-        selectedFields: [
-          config.dataFabricTestJoinFieldName,
-          `${relatedEntity}.${config.dataFabricTestJoinRelatedFieldName}`,
-        ],
-        joins: [
-          {
-            entityName: config.dataFabricTestJoinEntityName,
-            joinType: JoinType.LeftJoin,
-            joinFieldName: config.dataFabricTestJoinFieldName,
-            relatedEntityName: relatedEntity,
-            relatedFieldName: config.dataFabricTestJoinRelatedFieldName,
-          },
-        ],
-        pageSize: 25,
-      });
-
-      expect(Array.isArray(result.items)).toBe(true);
-      expect(result.items.length).toBeGreaterThan(0);
-      // Multi-entity result rows use entity-qualified keys ("Entity.Field").
-      // At least one base record has a matching related record, so at least one
-      // row must carry a key qualified with the related entity's name — this is
-      // the assertion that fails when the backend ignores the `joins` clause.
-      const joinedRows = result.items.filter(item =>
-        Object.keys(item).some(key => key.startsWith(`${relatedEntity}.`)),
-      );
-      expect(joinedRows.length).toBeGreaterThan(0);
-      expect(hasValidPagination(result)).toBe(true);
-    });
-
-    it('should return only matched rows for an INNER join', async () => {
-      const { entities } = getServices();
-      const config = getTestConfig();
-      const entityId = config.dataFabricTestEntityId || testEntityId;
-      if (!entityId) {
-        throw new Error('No entity ID available for testing');
-      }
-      if (
-        !config.dataFabricTestJoinFieldName ||
-        !config.dataFabricTestJoinRelatedEntityName ||
-        !config.dataFabricTestJoinRelatedFieldName
-      ) {
-        throw new Error('DATA_FABRIC_TEST_JOIN_* env vars are required for the join test');
-      }
-      const relatedEntity = config.dataFabricTestJoinRelatedEntityName;
-      const join = {
-        entityName: config.dataFabricTestJoinEntityName,
-        joinFieldName: config.dataFabricTestJoinFieldName,
-        relatedEntityName: relatedEntity,
-        relatedFieldName: config.dataFabricTestJoinRelatedFieldName,
-      };
-      const selectedFields = [
-        config.dataFabricTestJoinFieldName,
-        `${relatedEntity}.${config.dataFabricTestJoinRelatedFieldName}`,
-      ];
-
-      const left = await entities.queryRecordsById(entityId, {
-        selectedFields,
-        joins: [{ ...join, joinType: JoinType.LeftJoin }],
-        pageSize: 25,
-      });
-      const inner = await entities.queryRecordsById(entityId, {
-        selectedFields,
-        joins: [{ ...join, joinType: JoinType.InnerJoin }],
-        pageSize: 25,
-      });
-
-      // INNER keeps only matched rows; LEFT keeps every base row. The fixture
-      // guarantees at least one match, and every INNER row must carry the
-      // related entity's qualified keys.
-      expect(inner.items.length).toBeGreaterThan(0);
-      expect(inner.items.length).toBeLessThanOrEqual(left.items.length);
-      inner.items.forEach(item => {
-        expect(Object.keys(item).some(key => key.startsWith(`${relatedEntity}.`))).toBe(true);
-      });
-    });
-
-    it('should reject a join to a nonexistent entity', async () => {
-      const { entities } = getServices();
-      const config = getTestConfig();
-      const entityId = config.dataFabricTestEntityId || testEntityId;
-      if (!entityId) {
-        throw new Error('No entity ID available for testing');
-      }
-
-      // Negative control: the joins-aware route validates the join clause. If
-      // this resolves successfully, the SDK is hitting an endpoint that ignores
-      // `joins` — the regression this suite exists to prevent.
-      await expect(
-        entities.queryRecordsById(entityId, {
-          selectedFields: ['NoSuchEntityForJoinTest.Id'],
-          joins: [
-            {
-              joinType: JoinType.LeftJoin,
-              joinFieldName: 'Id',
-              relatedEntityName: 'NoSuchEntityForJoinTest',
-              relatedFieldName: 'Id',
-            },
-          ],
-          pageSize: 5,
-        }),
-      ).rejects.toThrow();
-    });
-  });
-
   describe('importRecordsById', () => {
     it('should import records from a CSV blob', async () => {
       const { entities } = getServices();
@@ -1103,526 +830,6 @@ describe.each(modes)('Data Fabric Entities - Integration Tests [%s]', (mode) => 
       expect(typeof result.insertedRecords).toBe('number');
       expect(result.totalRecords).toBeGreaterThanOrEqual(0);
     });
-  });
-
-
-  // ─── Schema Management ────────────────────────────────────────────────────
-
-  // Entity schema write operations require write-schema PAT scope — not supported yet
-  describe('create', () => {
-    it('should create a new entity and return its ID', async () => {
-      const { entities } = getServices();
-      const name = `sdk_test_${generateRandomString(8).toLowerCase()}`;
-
-      const entityId = await entities.create(name, [
-        { name: 'title', displayName: 'Title', type: EntityFieldDataType.STRING, isRequired: true },
-        { name: 'count', displayName: 'Count', type: EntityFieldDataType.DECIMAL, decimalPrecision: 0 },
-      ], { displayName: `SDK Test Entity ${name}`, description: 'Created by integration test' });
-
-      expect(typeof entityId).toBe('string');
-      expect(entityId.length).toBeGreaterThan(0);
-      createdEntityIds.push(entityId);
-    });
-
-    it('should create entity with RELATIONSHIP and FILE fields and verify metadata', async () => {
-      const { entities } = getServices();
-      const stamp = generateRandomString(8).toLowerCase();
-
-      // 1. Create the target entity. The platform auto-adds a primary-key `Id`
-      //    field we'll use as the FK target.
-      const targetName = `sdk_target_${stamp}`;
-      const targetId = await entities.create(targetName, [
-        { name: 'label', type: EntityFieldDataType.STRING },
-      ]);
-      createdEntityIds.push(targetId);
-
-      // 2. Find the primary-key field UUID on the target (referenceFieldId).
-      const targetMeta = await entities.getById(targetId);
-      const pkField = targetMeta.fields.find(f => f.isPrimaryKey);
-      if (!pkField?.id) {
-        throw new Error(`Target entity ${targetId} has no primary-key field — cannot bind a RELATIONSHIP to it`);
-      }
-
-      // 3. Create the source entity with a RELATIONSHIP field bound to target.Id
-      const sourceName = `sdk_source_${stamp}`;
-      const sourceId = await entities.create(sourceName, [
-        { name: 'name', type: EntityFieldDataType.STRING },
-        {
-          name: 'parent',
-          type: EntityFieldDataType.RELATIONSHIP,
-          referenceEntityId: targetId,
-          referenceFieldId: pkField.id,
-        },
-        { name: 'attachfile', type: EntityFieldDataType.FILE },
-      ]);
-      createdEntityIds.push(sourceId);
-
-      // 4. Read it back and verify both fields landed correctly. The server
-      //    resolves `referenceEntity { id }` → a full entity reference
-      const sourceMeta = await entities.getById(sourceId);
-      const parentField = sourceMeta.fields.find(f => f.name === 'parent');
-      expect(parentField).toBeDefined();
-      expect(parentField?.isForeignKey).toBe(true);
-      expect(parentField?.referenceEntity?.id).toBe(targetId);
-      expect(parentField?.referenceField?.id).toBe(pkField.id);
-
-      const fileField = sourceMeta.fields.find(f => f.name === 'attachfile');
-      expect(fileField).toBeDefined();
-      expect(fileField?.fieldDisplayType).toBe(FieldDisplayType.File);
-      // FILE fields are wired to the internal EntityAttachment blob-holder entity,
-      // not to a user-created reference — verify it's not pointing at our target.
-      expect(fileField?.referenceEntity?.name).toBe('EntityAttachment');
-      expect(fileField?.referenceEntity?.id).not.toBe(targetId);
-    }, 90_000);
-  });
-
-  describe('updateById', () => {
-    it('should update entity display name and description', async () => {
-      const { entities } = getServices();
-      const name = `sdk_test_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [], { displayName: `Original ${name}`, description: 'Original description' });
-      createdEntityIds.push(entityId);
-
-      const newDisplayName = `Updated ${name}`;
-      await entities.updateById(entityId, {
-        displayName: newDisplayName,
-        description: 'Updated description',
-      });
-
-      const updated = await entities.getById(entityId);
-      expect(updated.displayName).toBe(newDisplayName);
-      expect(updated.description).toBe('Updated description');
-    });
-
-    it('should add a new field to an existing entity', async () => {
-      const { entities } = getServices();
-      const name = `sdk_test_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'baseField', type: EntityFieldDataType.STRING },
-      ]);
-      createdEntityIds.push(entityId);
-
-      await entities.updateById(entityId, {
-        addFields: [{ name: 'newField', type: EntityFieldDataType.DECIMAL, decimalPrecision: 0 }],
-      });
-
-      const updated = await entities.getById(entityId);
-      const fieldNames = updated.fields.map(f => f.name);
-      expect(fieldNames).toContain('newField');
-    }, 60_000);
-
-    it('should remove a field from an existing entity', async () => {
-      const { entities } = getServices();
-      const name = `sdk_test_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'keepField', type: EntityFieldDataType.STRING },
-        { name: 'removeMe', type: EntityFieldDataType.DECIMAL, decimalPrecision: 0 },
-      ]);
-      createdEntityIds.push(entityId);
-
-      await entities.updateById(entityId, {
-        removeFields: [{ name: 'removeMe' }],
-      });
-
-      const updated = await entities.getById(entityId);
-      const fieldNames = updated.fields.map(f => f.name);
-      expect(fieldNames).not.toContain('removeMe');
-      expect(fieldNames).toContain('keepField');
-    });
-
-    it('should update an existing field metadata', async () => {
-      const { entities } = getServices();
-      const name = `sdk_test_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'updatableField', displayName: 'Original Name', type: EntityFieldDataType.STRING },
-      ]);
-      createdEntityIds.push(entityId);
-
-      // Get the raw entity to find the field ID (transformData renames sqlType but preserves id)
-      const before = await entities.getById(entityId);
-      const field = before.fields.find(f => f.name === 'updatableField');
-      if (!field?.id) {
-        throw new Error('Could not find updatable_field id in entity schema');
-      }
-
-      await entities.updateById(entityId, {
-        updateFields: [{ id: field.id, displayName: 'Updated Name', isRequired: true }],
-      });
-
-      const after = await entities.getById(entityId);
-      const updatedField = after.fields.find(f => f.name === 'updatableField');
-      expect(updatedField).toBeDefined();
-      expect(updatedField?.displayName).toBe('Updated Name');
-      expect(updatedField?.isRequired).toBe(true);
-    });
-
-    it('should add, update, and remove fields in a single call', async () => {
-      const { entities } = getServices();
-      const name = `sdk_test_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'toUpdate', displayName: 'Before Update', type: EntityFieldDataType.STRING },
-        { name: 'toRemove', type: EntityFieldDataType.DECIMAL, decimalPrecision: 0 },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const before = await entities.getById(entityId);
-      const fieldToUpdate = before.fields.find(f => f.name === 'toUpdate');
-      if (!fieldToUpdate?.id) {
-        throw new Error('Could not find to_update field id');
-      }
-
-      await entities.updateById(entityId, {
-        addFields: [{ name: 'newAddition', type: EntityFieldDataType.BOOLEAN }],
-        updateFields: [{ id: fieldToUpdate.id, displayName: 'After Update' }],
-        removeFields: [{ name: 'toRemove' }],
-      });
-
-      const after = await entities.getById(entityId);
-      const fieldNames = after.fields.map(f => f.name);
-
-      // new field was added
-      expect(fieldNames).toContain('newAddition');
-      // removed field is gone
-      expect(fieldNames).not.toContain('toRemove');
-      // updated field has new display name
-      const updated = after.fields.find(f => f.name === 'toUpdate');
-      expect(updated?.displayName).toBe('After Update');
-    });
-  });
-
-  describe('sqlType constraint defaults', () => {
-    it('should create STRING field with default lengthLimit 200', async () => {
-      const { entities } = getServices();
-      const name = `sdk_str_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'strField', type: EntityFieldDataType.STRING },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const entity = await entities.getById(entityId);
-      const field = entity.fields.find(f => f.name === 'strField');
-      expect(field).toBeDefined();
-      expect(field?.fieldDataType.lengthLimit).toBe(200);
-    });
-
-    it('should create STRING field with user-provided lengthLimit', async () => {
-      const { entities } = getServices();
-      const name = `sdk_str_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'strField', type: EntityFieldDataType.STRING, lengthLimit: 500 },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const entity = await entities.getById(entityId);
-      const field = entity.fields.find(f => f.name === 'strField');
-      expect(field?.fieldDataType.lengthLimit).toBe(500);
-    });
-
-    it('should create MULTILINE_TEXT field with default lengthLimit 200', async () => {
-      const { entities } = getServices();
-      const name = `sdk_ml_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'mlField', type: EntityFieldDataType.MULTILINE_TEXT },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const entity = await entities.getById(entityId);
-      const field = entity.fields.find(f => f.name === 'mlField');
-      expect(field?.fieldDataType.lengthLimit).toBe(200);
-    });
-
-    it('should create MULTILINE_MAX field with default lengthLimit 128 KB', async () => {
-      const { entities } = getServices();
-      const name = `sdk_mlmax_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'mlmaxField', type: EntityFieldDataType.MULTILINE_MAX },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const entity = await entities.getById(entityId);
-      const field = entity.fields.find(f => f.name === 'mlmaxField');
-      expect(field?.fieldDataType.name).toBe(EntityFieldDataType.MULTILINE_MAX);
-      expect(field?.fieldDataType.lengthLimit).toBe(128 * 1024);
-    });
-
-    it('should create DECIMAL field with correct default constraints', async () => {
-      const { entities } = getServices();
-      const name = `sdk_dec_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'decField', type: EntityFieldDataType.DECIMAL },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const entity = await entities.getById(entityId);
-      const field = entity.fields.find(f => f.name === 'decField');
-      expect(field?.fieldDataType.lengthLimit).toBe(1000);
-      expect(field?.fieldDataType.decimalPrecision).toBe(2);
-      expect(field?.fieldDataType.maxValue).toBe(1000000000000);
-      expect(field?.fieldDataType.minValue).toBe(-1000000000000);
-    });
-
-    it('should create DECIMAL field with user-provided constraints', async () => {
-      const { entities } = getServices();
-      const name = `sdk_dec_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        {
-          name: 'decField',
-          type: EntityFieldDataType.DECIMAL,
-          decimalPrecision: 4,
-          maxValue: 99999,
-          minValue: -99999,
-        },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const entity = await entities.getById(entityId);
-      const field = entity.fields.find(f => f.name === 'decField');
-      expect(field?.fieldDataType.decimalPrecision).toBe(4);
-      expect(field?.fieldDataType.maxValue).toBe(99999);
-      expect(field?.fieldDataType.minValue).toBe(-99999);
-    });
-
-    it('should create BOOLEAN field with fixed lengthLimit 100', async () => {
-      const { entities } = getServices();
-      const name = `sdk_bit_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'boolField', type: EntityFieldDataType.BOOLEAN },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const entity = await entities.getById(entityId);
-      const field = entity.fields.find(f => f.name === 'boolField');
-      expect(field?.fieldDataType.lengthLimit).toBe(100);
-    });
-
-    it('should create DATE field with fixed lengthLimit 1000', async () => {
-      const { entities } = getServices();
-      const name = `sdk_date_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'dateField', type: EntityFieldDataType.DATE },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const entity = await entities.getById(entityId);
-      const field = entity.fields.find(f => f.name === 'dateField');
-      expect(field?.fieldDataType.lengthLimit).toBe(1000);
-    });
-
-    it('should create DATETIME_WITH_TZ field with fixed lengthLimit 1000', async () => {
-      const { entities } = getServices();
-      const name = `sdk_dtz_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'dtzField', type: EntityFieldDataType.DATETIME_WITH_TZ },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const entity = await entities.getById(entityId);
-      const field = entity.fields.find(f => f.name === 'dtzField');
-      expect(field?.fieldDataType.lengthLimit).toBe(1000);
-    });
-
-    it('should allow updating STRING field lengthLimit without "Field type cannot be changed" error', async () => {
-      const { entities } = getServices();
-      const name = `sdk_upd_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'strField', type: EntityFieldDataType.STRING },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const before = await entities.getById(entityId);
-      const field = before.fields.find(f => f.name === 'strField');
-      if (!field?.id) {
-        throw new Error('Could not find str_field id in entity schema');
-      }
-
-      // Must not throw "Field type cannot be changed"
-      await entities.updateById(entityId, {
-        updateFields: [{ id: field.id, lengthLimit: 500 }],
-      });
-
-      const after = await entities.getById(entityId);
-      const updated = after.fields.find(f => f.name === 'strField');
-      expect(updated?.fieldDataType.lengthLimit).toBe(500);
-    }, 60_000);
-
-    it('should allow updating DECIMAL field constraints via updateById', async () => {
-      const { entities } = getServices();
-      const name = `sdk_upddec_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, [
-        { name: 'decField', type: EntityFieldDataType.DECIMAL },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const before = await entities.getById(entityId);
-      const field = before.fields.find(f => f.name === 'decField');
-      if (!field?.id) {
-        throw new Error('Could not find dec_field id in entity schema');
-      }
-
-      await entities.updateById(entityId, {
-        updateFields: [{ id: field.id, decimalPrecision: 4, maxValue: 9999, minValue: -9999 }],
-      });
-
-      const after = await entities.getById(entityId);
-      const updated = after.fields.find(f => f.name === 'decField');
-      expect(updated?.fieldDataType.decimalPrecision).toBe(4);
-      expect(updated?.fieldDataType.maxValue).toBe(9999);
-      expect(updated?.fieldDataType.minValue).toBe(-9999);
-    }, 60_000);
-
-    it('should apply default lengthLimit (200) when STRING lengthLimit is omitted, confirmed via GET', async () => {
-      const { entities } = getServices();
-      const name = `sdk_str_default_${generateRandomString(8).toLowerCase()}`;
-
-      const entityId = await entities.create(name, [
-        { name: 'strField', type: EntityFieldDataType.STRING },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const entity = await entities.getById(entityId);
-      const field = entity.fields.find(f => f.name === 'strField');
-      // Default is applied client-side and round-trips through the API as 200
-      expect(field?.fieldDataType.lengthLimit).toBe(200);
-    }, 60_000);
-
-  });
-
-  // Custom user-defined RELATIONSHIP fields must follow the same expansion rules as
-  // system reference fields. Placed after sqlType so both modes hit a warm DF backend
-  // (avoids read-after-write flakes on the first schema-write of the session).
-  describe('RELATIONSHIP field expansion', () => {
-    it('should expand a custom RELATIONSHIP field at each expansionLevel (0-3)', async () => {
-      const { entities } = getServices();
-      const stamp = generateRandomString(8).toLowerCase();
-
-      // 1. Target entity with a user-defined string field we can assert on at L2+.
-      const targetId = await entities.create(`sdk_target_${stamp}`, [
-        { name: 'label', type: EntityFieldDataType.STRING, isRequired: true },
-      ]);
-      createdEntityIds.push(targetId);
-
-      const targetMeta = await entities.getById(targetId);
-      const targetPk = targetMeta.fields.find(f => f.isPrimaryKey);
-      if (!targetPk?.id) {
-        throw new Error(`Target entity ${targetId} has no primary-key field`);
-      }
-
-      // 2. Source entity with a RELATIONSHIP field bound to target.Id.
-      const sourceId = await entities.create(`sdk_source_${stamp}`, [
-        { name: 'name', type: EntityFieldDataType.STRING },
-        {
-          name: 'parent',
-          type: EntityFieldDataType.RELATIONSHIP,
-          referenceEntityId: targetId,
-          referenceFieldId: targetPk.id,
-        },
-      ]);
-      createdEntityIds.push(sourceId);
-
-      // 3. Insert a target record so the FK has something to resolve.
-      const labelValue = `Target_${stamp}`;
-      const targetInsert = await entities.insertRecordById(targetId, { label: labelValue });
-      const targetRecordId = targetInsert.Id;
-      registerResource('entityRecords', { entityId: targetId, recordIds: [targetRecordId] });
-
-      // 4. Insert the source record with the FK populated.
-      const sourceInsert = await entities.insertRecordById(sourceId, {
-        name: `Source_${stamp}`,
-        parent: targetRecordId,
-      });
-      registerResource('entityRecords', { entityId: sourceId, recordIds: [sourceInsert.Id] });
-
-      // 5. Query the source at every expansion level.
-      const levels = [0, 1, 2, 3] as const;
-      const responses = await Promise.all(
-        levels.map(level => entities.queryRecordsById(sourceId, { expansionLevel: level, pageSize: 10 })),
-      );
-
-      const sourceRecords = responses.map(r =>
-        (r.items as Record<string, any>[]).find(item => item.Id === sourceInsert.Id),
-      );
-      sourceRecords.forEach((rec, i) => {
-        expect(rec, `expansionLevel=${levels[i]} did not return the inserted source record`).toBeDefined();
-      });
-      const [l0, l1, l2, l3] = sourceRecords as Record<string, any>[];
-
-      // L0: FK is the raw target record Id string.
-      expect(typeof l0.parent).toBe('string');
-      expect(l0.parent).toBe(targetRecordId);
-
-      // L1+: FK inflates into an object envelope carrying the target Id.
-      for (const [level, rec] of [[1, l1], [2, l2], [3, l3]] as const) {
-        expect(typeof rec.parent, `L${level} parent should be object`).toBe('object');
-        expect(rec.parent, `L${level} parent should not be null`).not.toBeNull();
-        expect(rec.parent, `L${level} parent should carry target Id`).toHaveProperty('Id', targetRecordId);
-      }
-
-      // L2 surfaces the user-defined `label` field from the target record.
-      expect(l2.parent.label).toBe(labelValue);
-
-      // L3 cannot shrink relative to L2.
-      expect(Object.keys(l3.parent).length).toBeGreaterThanOrEqual(Object.keys(l2.parent).length);
-    }, 150_000);
-  });
-
-  // Verifies the MULTILINE_MAX round-trip end to end: create + insert + read the
-  // full value back via getRecordById (v2 read). The list endpoint's size-marker
-  // projection is not asserted — the server threshold is inconsistent and the
-  // consumer-facing guarantee is the full-content read.
-  describe('MULTILINE_MAX field lifecycle', () => {
-    it('should round-trip a MULTILINE_MAX value via getRecordById', async () => {
-      const { entities } = getServices();
-      const name = `sdk_mlmax_life_${generateRandomString(8).toLowerCase()}`;
-
-      const entityId = await entities.create(name, [
-        { name: 'body', type: EntityFieldDataType.MULTILINE_MAX },
-      ]);
-      createdEntityIds.push(entityId);
-
-      const bodyValue = `Large body content ${generateRandomString(256)}`;
-      const inserted = await entities.insertRecordById(entityId, { body: bodyValue });
-      expect(inserted.Id).toBeDefined();
-      registerResource('entityRecords', { entityId, recordIds: [inserted.Id] });
-
-      const full = await entities.getRecordById(entityId, inserted.Id);
-      expect(full.body).toBe(bodyValue);
-    }, 90_000);
-  });
-
-  describe('deleteById', () => {
-    it('should delete an entity created for this test', async () => {
-      const { entities } = getServices();
-
-      const name = `sdk_test_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, []);
-
-      // Delete it immediately (not added to createdEntityIds so afterAll won't double-delete)
-      await entities.deleteById(entityId);
-
-      // Verify it no longer appears in getAll
-      const all = await entities.getAll();
-      const found = all.find(e => e.id === entityId);
-      expect(found).toBeUndefined();
-    });
-  });
-
-  describe('entity schema methods (bound)', () => {
-    it('should call deleteById via bound method on entity', async () => {
-      const { entities } = getServices();
-
-      const name = `sdk_test_${generateRandomString(8).toLowerCase()}`;
-      const entityId = await entities.create(name, []);
-
-      const entity = await entities.getById(entityId);
-      expect(typeof entity.delete).toBe('function');
-      expect(typeof entity.update).toBe('function');
-
-      // Delete via bound method — do NOT add to createdEntityIds
-      await entity.delete();
-
-      const all = await entities.getAll();
-      expect(all.find(e => e.id === entityId)).toBeUndefined();
-    }, 60_000);
   });
 
   // ─── Single Record Delete ─────────────────────────────────────────────────
@@ -1869,17 +1076,6 @@ describe.each(modes)('Data Fabric Entities - Integration Tests [%s]', (mode) => 
     if (!config.skipCleanup) {
       if (createdRecordIds.length > 0 && testEntityId) {
         await cleanupTestEntityRecords(testEntityId, createdRecordIds);
-      }
-      // Clean up any entities created by schema management tests (no-op when those tests are skipped)
-      if (createdEntityIds.length > 0) {
-        const { entities } = getServices();
-        await Promise.all(
-          createdEntityIds.map(entityId =>
-            entities.deleteById(entityId).catch(() => {
-              // Best-effort cleanup — entity may have already been deleted
-            })
-          )
-        );
       }
     }
   });
